requirements.txt
Flask==0.11.1
twilio==5.6.0
PyYAML==3.12
paho-mqtt>=1.2
